+++ tests/e2e/core-tests/specs/merchant/wp-admin-settings-shipping-zones.test.js
@@ -10,8 +10,7 @@ const {
 	addShippingZoneAndMethod,
 	clearAndFillInput,
 	selectOptionInSelect2,
-	evalAndClick,
-	uiUnblocked,
+	deleteAllShippingZones,
 } = require( '@woocommerce/e2e-utils' );
 
 const config = require( 'config' );
@@ -28,27 +27,7 @@ const runAddNewShippingZoneTest = () => {
 		beforeAll(async () => {
 			await merchant.login();
 			await createSimpleProduct();
-			await merchant.openSettings('shipping');
-
-			// Delete existing shipping zones.
-			try {
-				let zone = await page.$( '.wc-shipping-zone-delete' );
-				if ( zone ) {
-					// WP action links aren't clickable because they are hidden with a left=-9999 style.
-					await page.evaluate(() => {
-						document.querySelector('.wc-shipping-zone-name .row-actions')
-							.style
-							.left = '0';
-					});
-					while ( zone ) {
-						await evalAndClick( '.wc-shipping-zone-delete' );
-						await uiUnblocked();
-						zone = await page.$( '.wc-shipping-zone-delete' );
-					}
-				}
-			} catch (error) {
-				// Prevent an error here causing the test to fail.
-			}
+			await deleteAllShippingZones();
 		});
 
 		it('add shipping zone for San Francisco with free Local pickup', async () => {