+++ tests/e2e/utils/src/components.js
@@ -12,7 +12,8 @@ import {
 	verifyCheckboxIsUnset,
 	selectOptionInSelect2,
 	setCheckbox,
-	unsetCheckbox
+	unsetCheckbox,
+	evalAndClick,
 } from './page-utils';
 import factories from './factories';
 
@@ -538,6 +539,32 @@ const deleteAllEmailLogs = async () => {
 	}
 };
 
+/**
+ * Delete all the existing shipping zones.
+ */
+const deleteAllShippingZones = async () => {
+	await merchant.openSettings('shipping');
+	// Delete existing shipping zones.
+	try {
+		let zone = await page.$( '.wc-shipping-zone-delete' );
+		if ( zone ) {
+			// WP action links aren't clickable because they are hidden with a left=-9999 style.
+			await page.evaluate(() => {
+				document.querySelector('.wc-shipping-zone-name .row-actions')
+					.style
+					.left = '0';
+			});
+			while ( zone ) {
+				await evalAndClick( '.wc-shipping-zone-delete' );
+				await uiUnblocked();
+				zone = await page.$( '.wc-shipping-zone-delete' );
+			};
+		};
+	} catch (error) {
+		// Prevent an error here causing the test to fail.
+	};
+};
+
 export {
 	completeOnboardingWizard,
 	createSimpleProduct,
@@ -551,4 +578,5 @@ export {
 	createSimpleProductWithCategory,
 	clickUpdateOrder,
 	deleteAllEmailLogs,
+	deleteAllShippingZones,
 };